requirements.txt
streamlit==1.35.0
opencv-python
mediapipe
PyYAML
numpy
Pillow
ipykernel